extracted_projects/Maruthi0302-Group-L-08856c2/requirements.txt
langchain
langchain-core
langchain-community
langgraph
langchain-groq
chromadb
sentence-transformers
fastapi
uvicorn
pydantic
pandas
kaggle
python-dotenv
loguru
requests
